requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
flask>=3.0.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0
